Move models to links/model & Make subdir of links
