+++ chainer_chemistry/__init__.py
@@ -12,7 +12,6 @@ except ImportError:
         UserWarning)
 from chainer_chemistry import functions  # NOQA
 from chainer_chemistry import links  # NOQA
-from chainer_chemistry import models  # NOQA
 from chainer_chemistry import training  # NOQA
 
 # --- config variable definitions ---
