+++ chainer_chemistry/links/__init__.py
@@ -1,6 +1,15 @@
-from chainer_chemistry.links import embed_atom_id  # NOQA
-from chainer_chemistry.links import graph_linear  # NOQA
+from chainer_chemistry.links.connection.embed_atom_id import EmbedAtomID  # NOQA
+from chainer_chemistry.links.connection.graph_linear import GraphLinear  # NOQA
 
-from chainer_chemistry.links.embed_atom_id import EmbedAtomID  # NOQA
-from chainer_chemistry.links.graph_batch_normalization import GraphBatchNormalization  # NOQA
-from chainer_chemistry.links.graph_linear import GraphLinear  # NOQA
+from chainer_chemistry.links.model.prediction.base import BaseForwardModel  # NOQA
+from chainer_chemistry.links.model.prediction.classifier import Classifier  # NOQA
+from chainer_chemistry.links.model.prediction.regressor import Regressor  # NOQA
+
+from chainer_chemistry.links.model.ggnn import GGNN  # NOQA
+from chainer_chemistry.links.model.mlp import MLP  # NOQA
+from chainer_chemistry.links.model.nfp import NFP  # NOQA
+from chainer_chemistry.links.model.rsgcn import RSGCN  # NOQA
+from chainer_chemistry.links.model.schnet import SchNet  # NOQA
+from chainer_chemistry.links.model.weavenet import WeaveNet  # NOQA
+
+from chainer_chemistry.links.normalization.graph_batch_normalization import GraphBatchNormalization  # NOQA
