+++ chainer_chemistry/links/connection/embed_atom_id.py
