+++ chainer_chemistry/links/connection/graph_linear.py
