+++ chainer_chemistry/links/model/ggnn.py
