+++ chainer_chemistry/links/model/nfp.py
