+++ chainer_chemistry/links/model/prediction/base.py
