+++ chainer_chemistry/links/model/prediction/classifier.py
