+++ chainer_chemistry/links/model/weavenet.py
@@ -4,7 +4,7 @@ from chainer import links
 
 from chainer_chemistry.config import MAX_ATOMIC_NUM
 from chainer_chemistry.config import WEAVE_DEFAULT_NUM_MAX_ATOMS
-from chainer_chemistry.links.embed_atom_id import EmbedAtomID
+from chainer_chemistry.links import EmbedAtomID
 
 
 WEAVENET_DEFAULT_WEAVE_CHANNELS = [50, ]
