+++ tests/links_tests/model_tests/test_schnet.py
@@ -4,7 +4,7 @@ import numpy
 import pytest
 
 from chainer_chemistry.config import MAX_ATOMIC_NUM
-from chainer_chemistry.models.nfp import NFP
+from chainer_chemistry.links.model.schnet import SchNet
 from chainer_chemistry.utils.permutation import permute_adj
 from chainer_chemistry.utils.permutation import permute_node
 
@@ -15,7 +15,7 @@ batch_size = 2
 
 @pytest.fixture
 def model():
-    return NFP(out_dim=out_dim)
+    return SchNet(out_dim=out_dim)
 
 
 @pytest.fixture
@@ -24,9 +24,12 @@ def data():
     atom_data = numpy.random.randint(
         0, high=MAX_ATOMIC_NUM, size=(batch_size, atom_size)
     ).astype(numpy.int32)
-    adj_data = numpy.random.randint(
-        0, high=2, size=(batch_size, atom_size, atom_size)
+    # symmetric matrix
+    adj_data = numpy.random.uniform(
+        0, high=30, size=(batch_size, atom_size, atom_size)
     ).astype(numpy.float32)
+    adj_data = (adj_data + adj_data.swapaxes(-1, -2)) / 2.
+
     y_grad = numpy.random.uniform(
         -1, 1, (batch_size, out_dim)).astype(numpy.float32)
     return atom_data, adj_data, y_grad
@@ -49,20 +52,18 @@ def test_forward_gpu(model, data):
     check_forward(model, atom_data, adj_data)
 
 
-# TODO(nakago): check why tolerance is high
 def test_backward_cpu(model, data):
     atom_data, adj_data, y_grad = data
     gradient_check.check_backward(model, (atom_data, adj_data), y_grad,
-                                  atol=1e0, rtol=1e0)
+                                  atol=5e-1, rtol=1e-1)
 
 
-# TODO(nakago): check why tolerance is high
 @pytest.mark.gpu
 def test_backward_gpu(model, data):
     atom_data, adj_data, y_grad = [cuda.to_gpu(d) for d in data]
     model.to_gpu()
     gradient_check.check_backward(model, (atom_data, adj_data), y_grad,
-                                  atol=1e0, rtol=1e0)
+                                  atol=5e-1, rtol=1e-1)
 
 
 def test_forward_cpu_graph_invariant(model, data):
@@ -74,7 +75,7 @@ def test_forward_cpu_graph_invariant(model, data):
     permute_adj_data = permute_adj(adj_data, permutation_index)
     permute_y_actual = cuda.to_cpu(model(
         permute_atom_data, permute_adj_data).data)
-    assert numpy.allclose(y_actual, permute_y_actual, rtol=1e-5, atol=1e-6)
+    assert numpy.allclose(y_actual, permute_y_actual, rtol=1e-5, atol=1e-5)
 
 
 if __name__ == '__main__':
